Add DRM mode stereo page-flip option to test API.
The drmModePageFlip() API was extended with an additional
'target' argument for VSync behavior, so overload the extra
argument to handle stereo page-flip offset for the right-eye
framebuffer when denoted with additional flag:

	DRM_MODE_PAGE_FLIP_TARGET_STEREO

For this case, the 'target' argument refers to the offset
between the left and right framebuffers for loading into
the corresponding pair of display start registers.
VSync behavior defaults to immediate async mode anyway
when stereo page-flipping mode is active since the display
controller is automatically alternating between left/right
output fields.

TODO: Since this is just proof-of-concept for the extra
stereo offset argument to be propagated through the DRM
driver stack, it still needs to call into the specific
Radeon chipset series driver code for R100, R200, etc.